fix: provide valid JWTs to credentials tests
