+++ tests/functional/multitenant/mod.rs
@@ -22,3 +22,23 @@ async fn test_health(ctx: &mut EchoServerContext) {
         .status();
     assert!(body.is_success());
 }
+
+pub fn generate_random_tenant_id() -> (String, String) {
+    let charset = "1234567890";
+    let random_tenant_id = generate(12, charset);
+    let unix_timestamp = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap()
+        .as_secs() as usize;
+    let token_claims = ClaimsForValidation {
+        sub: random_tenant_id.clone(),
+        exp: unix_timestamp + 60 * 60, // Add an hour for expiration
+    };
+    let jwt_token = encode(
+        &Header::default(),
+        &token_claims,
+        &EncodingKey::from_secret(ctx.config.jwt_secret.as_bytes()),
+    )
+    .expect("Failed to encode jwt token");
+    (random_tenant_id, jwt_token)
+}
